Split router into modular files
