+++ app/api_v1/player.py
@@ -1,4 +1,3 @@
-import os
 from datetime import datetime, timedelta
 from fastapi import APIRouter, HTTPException
 from pymongo.errors import PyMongoError, DuplicateKeyError
@@ -9,46 +8,6 @@ from app.includes.Requests import AuthenticateRequest
 from app.includes.Hash import ncchash, create_access_token
 from ..db.config import ACCESS_TOKEN_EXPIRE_MINUTES
 router = APIRouter()
-
-'''
-General API Status
-'''
-# Get health status of the API
-@router.get("/", tags=["Status"])
-def health_check():
-  return {"Status": "Running"}
-
-# Get all player scores in descending order
-@router.get("/get-scoreboard", tags=["Status"])
-async def get_scoreboard():
-  try:
-    users = players_collection.find({}, {"_id": 0, "username": 1, "score": 1}).sort("score", -1)
-    return [{"username": user["username"], "score": user["score"]} for user in users]
-  except PyMongoError as e:
-    raise HTTPException(status_code=500, detail=f"Database error: {e}")
-
-
-'''
-Admin Access
-'''
-# Add player score
-@router.put("/add-extra-credit", tags=["Admin"])
-# async def update_score(nccid: str, score: int, password: str = Depends(verify_admin_password)):
-async def add_points(nccid: str, score: int):
-  try:
-    # TODO: Review score update and extra credit list
-    players_collection.find_one_and_update(
-      {"nccid": nccid},
-      {
-        "$inc": {"score": score},
-        "$set": {"updatedAt": datetime.now()}
-      },
-      return_document=True
-    )
-    return {"message": "Score updated successfully", "nccid": nccid, "score": score}
-  except Exception as e:
-    raise HTTPException(status_code=500, detail=str(e))
-
 
 '''
 Player Profile Management
